check for existing oscar.war

+++ build-source.sh
@@ -7,7 +7,10 @@ echo "This deploys a fresh oscar from source."
 echo "Compiling OSCAR. This may take some time...."
 docker-compose -f docker-compose.admin.yml run builder ./bin/build-oscar.sh
 
-cp ./oscar/target/oscar-*-SNAPSHOT.war oscar.war
+if [ ! -f oscar.war ]; then
+  echo "Building oscar.war from cloned Oscar REPO"
+  cp ./oscar/target/oscar-*-SNAPSHOT.war oscar.war
+fi
 
 export DRUGREF_WAR=${DRUGREF_WAR:-"http://bool.countable.ca/drugref2.war"}
